refactor(imports): change all require calls to imports of module relays
