+++ src/errors/base-error/base-error.test.ts
@@ -1,19 +1,16 @@
+import { BaseError } from '.';
+
 describe('Module ts-utils/errors/base-error:', () => {
     test('should export a Class BaseError', () => {
-        const { BaseError } = require('./base-error');
-
         expect(BaseError).toBeInstanceOf(Function);
     });
-    
+
     describe('Exported class BaseError:', () => {
         test('should be a subclass of Error class', () => {
-            const { BaseError } = require('./base-error');
-
             expect(BaseError.prototype).toBeInstanceOf(Error);
         });
 
         test('should be a throwable', () => {
-            const { BaseError } = require('./base-error');
             const testErrorMessage = 'This is no error!!?';
 
             try {
