+++ src/mandate/mandate.test.ts
@@ -1,22 +1,17 @@
-import { MissingMandatoryParamError } from './../errors';
+import { MissingMandatoryParamError } from './../errors/missing-mandatory-param-error';
+import { mandate } from '.';
 
 describe('Module ts-utils/mandate:', () => {
     test('should export a function "mandate"', () => {
-        const { mandate } = require('.');
-
         expect(mandate).toBeInstanceOf(Function);
     });
 
     describe('Exported function "mandate":', () => {
         test('should throw a MissingMandatoryParamError when called', () => {
-            const { mandate } = require('.');
-
             expect(mandate).toThrowError(MissingMandatoryParamError);
         });
 
         test('should specify the paramName passed in the MissingMandatoryParamError thrown', () => {
-            const { mandate } = require('.');
-
             expect(() => mandate('someParam')).toThrowError(/.*"someParam"/);
         });
     });
