+++ src/mandate/mandate.ts
@@ -1,4 +1,4 @@
-import { MissingMandatoryParamError } from './../errors';
+import { MissingMandatoryParamError } from './../errors/missing-mandatory-param-error';
 
 export function mandate(paramName: string): any {
     throw new MissingMandatoryParamError(paramName);
